user homepage bugs
