+++ src/components/ApplicationViews.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Route } from "react-router-dom"
-import { Profile } from "./user/Profile.js"
+import { Profile } from "./Profile.js"
 import { ProfileProvider } from "./auth/ProfileProvider.js"
 import { Home } from "./Home.js"
 
@@ -10,17 +10,20 @@ export const ApplicationViews = () => {
             margin: "5rem 2rem",
             lineHeight: "1.75rem"
         }}>
+            <ProfileProvider>
             
-            <Route exact path="/">
-                <Home/>
-            </Route>
-    
-            {/* ______________posts__________________________ */}
+                <Route exact path="/">
+                    <Home/>
+                </Route>
+        
+                {/* ______________posts__________________________ */}
 
-            {/* _________________Profile______________________________ */}
-            <Route exact path="/profile">
-                <Profile />
-            </Route>
+                {/* _________________Profile______________________________ */}
+                <Route exact path="/profile">
+                    <Profile />
+                </Route>
+                
+            </ProfileProvider>
         </main>
     </>
 }
