+++ src/components/auth/Register.js
@@ -36,7 +36,7 @@ export const Register = (props) => {
                 .then(res => {
                     if ("token" in res) {
                         localStorage.setItem("lu_token", res.token)
-                        history.push("/")
+                        history.push("/profile")
                     }
                 })
         } else {
